Add files via upload
A To-Do List is a simple task management tool that helps users keep track of tasks they need to complete. It is a list where tasks can be added, marked as complete, and removed once finished. Typically, it includes features like adding, updating, deleting, and categorizing tasks based on priority or deadlines.
Just Before runing code do npm init so that respectve node module gets generated into the system. then do npm start so that to do list web application will be running on the browser.